Fix some type conversion warnings

+++ system/cpu.cc
@@ -76,7 +76,8 @@ void Cpu::RunSingleIteration(Graphics* graphics, Input* input, Memory* memory,
 
 std::uint16_t Cpu::GetCurrentOpcode(Memory* memory) {
   return (static_cast<std::uint16_t>(memory->ReadByte(pc_)) << 8) |
-         static_cast<std::uint16_t>(memory->ReadByte(pc_ + 1));
+      static_cast<std::uint16_t>(
+          memory->ReadByte(static_cast<std::size_t>(pc_) + 1));
 }
 
 void Cpu::ExecuteOpcode(std::uint16_t opcode, Graphics* graphics, Input* input,
@@ -400,7 +401,7 @@ void Cpu::StoreBinaryCodedDecimalAtVxInMemory(std::uint16_t opcode,
       static_cast<std::size_t>(v_registers_[decode_x(opcode)]);
 
   for (int i = 2; i >= 0; i--) {
-    memory->WriteByte(i_register_ + i, binary_coded_decimal & 1);
+    memory->WriteByte(static_cast<std::size_t>(i_register_) + i, binary_coded_decimal & 1);
     binary_coded_decimal = binary_coded_decimal >> 1;
   }
 }
